Fix Domain composition for greater than 2 billion particles.

+++ Sorter.h
@@ -35,7 +35,7 @@ struct OctDecompNode {
   void makeSubTree(int refineLevel, CkVec<OctDecompNode*> *active);
   int buildCounts();
   void deleteBeneath();
-  void combine(int thresh, vector<NodeKey> &finalKeys, vector<unsigned int> &counts);
+  void combine(int thresh, vector<NodeKey> &finalKeys, vector<uint64_t> &counts);
   void getLeafNodes(CkVec<OctDecompNode*> *activeNodes);
 };
 
@@ -65,11 +65,9 @@ class Sorter : public CBase_Sorter {
 
         double decompTime;
 	/// The total number of keys we're sorting.
-	int numKeys;
+	int64_t numKeys;
 	/// The number of chares to sort into.
 	int numChares;
-	// The number of chares currently with assigned data.
-	//int numUsedChares;
 	/// The indexes of the chares that are responsible for each segment of data.
 	std::vector<int> chareIDs;
 	/// A list of chare elements to which nothing is assigned
@@ -88,7 +86,7 @@ class Sorter : public CBase_Sorter {
 
 	std::vector<NodeKey> nodeKeys;
 	/// The histogram of counts for the last round of splitter keys.
-	std::vector<unsigned int> binCounts;
+	std::vector<uint64_t> binCounts;
 	std::vector<unsigned int> binCountsGas;
 	std::vector<unsigned int> binCountsStar;
 	/// The number of bins in the histogram.
@@ -101,7 +99,7 @@ class Sorter : public CBase_Sorter {
 
         CkBitVector binsToSplit;
 	/// The list of object number splits not yet met.
-	int *goals;
+	int64_t *goals;
         int numGoalsPending;
 	
 	/// The DataManager I broadcast candidate keys to.
